Initial Early Test Concept using vBoxSysInfoMod v2
Taking and using vBoxSysInfoMod v2 aka the bugged non-working version that messes up or bricks any and all vBox VMs that you try to use/run it on that was kept purely and solely for archival purposes, and using it to create a very early sample and concept to repairing/fixing and/or resetting any and all changes made by vBoxSysInfoMod, with the exception of the paravirtualization interface change which was added back in vBoxSysInfoMod v4 which hides the VM indicators from various locations such as Task Manager. Although this is more or less just an alpha, it is stable and works as is although it is based off of a modified version of vBoxSysInfoMod v2 which is pretty much very old and outdated by the current standard which is vBoxSysInfoMod v5.2 at the time of writing so this will be updated, cleaned and optimized whenever I get the chance.